requirements.txt
Django==5.2.6
channels==4.0.0
channels-redis==4.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
panoramisk==1.4
asyncio-mqtt==0.16.2
cryptography==43.0.1
django-redis==5.4.0
nvidia-riva-client==2.19.0
soundfile==0.13.1
numpy==2.3.3
scipy==1.16.2
pyrnnoise==0.3.8
webrtcvad-wheels==2.0.14
nemo_toolkit[asr]
torch==2.8.0
torchaudio==2.8.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
drf-spectacular==0.26.5
django-cors-headers==4.3.1
openpyxl==3.1.2
